% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey_stroke.R
\name{plot_sankey_stroke}
\alias{plot_sankey_stroke}
\title{Plot Sankey diagram for stroke risk contributors}
\usage{
plot_sankey_stroke(
  df,
  plot_title = "Contributors to Stroke in Blacks",
  vertical_gap = 0.02,
  horizontal_gap = 0,
  block_width = 0.7,
  min_h = 0.05,
  smooth_pts = 40
)
}
\arguments{
\item{df}{Dataset containing stroke PAR values. Can be:
\enumerate{
\item wide raw Excel-style data,
\item wide cleaned data with RiskFactor + year columns,
\item long-format data with RiskFactor, Year, PAR.
}}

\item{plot_title}{Title of the plot}

\item{vertical_gap}{Vertical whitespace between ribbons}

\item{horizontal_gap}{Horizontal whitespace between year blocks}

\item{block_width}{Width of the year blocks}

\item{min_h}{Minimum visual ribbon height}

\item{smooth_pts}{Number of interpolation points}
}
\description{
Plot Sankey diagram for stroke risk contributors
}
